Attempt to fix issue r3458. I don't have a 32bit linux install, however I know it will at least fix compressing. Please test if uncompressed games run.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@6510 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/DiscIO/Src/VolumeGC.cpp
@@ -129,7 +129,7 @@ std::string CVolumeGC::GetApploaderDate() const
 u64 CVolumeGC::GetSize() const
 {
 	if (m_pReader)
-		return (size_t)m_pReader->GetDataSize();
+		return m_pReader->GetDataSize();
 	else
 		return 0;
 }
